requirements.txt
reflex>=0.6.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
